Cascade storybook adjustments
- configure cascade multiple stats story to render 100K rows

- split data cascade stories into managable chunks
