+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/__stories__/scenes/row_selection.tsx
@@ -1,0 +1,375 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import {
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiIcon,
+  EuiText,
+  EuiFilterGroup,
+  EuiFilterButton,
+  EuiButtonGroup,
+  EuiStat,
+  EuiHealth,
+  EuiBasicTable,
+  type HorizontalAlignment,
+} from '@elastic/eui';
+import { faker } from '@faker-js/faker';
+import { getESQLStatsQueryMeta } from '@kbn/esql-utils';
+import type { StoryObj } from '@storybook/react';
+import React, { type ComponentProps, useMemo, useCallback } from 'react';
+import { FormattedMessage } from '@kbn/i18n-react';
+import { type MockGroupData } from '../../__fixtures__/types';
+import {
+  DataCascade,
+  type DataCascadeRowProps,
+  type DataCascadeRowCellProps,
+  DataCascadeRow,
+  DataCascadeRowCell,
+  type LeafNode,
+} from '../../components';
+
+/**
+ * This story demonstrates a custom header with row selection action enabled,
+ * this is useful for scenarios where one might want to select rows in the cascade component
+ * and perform actions on the selected rows.
+ */
+export const CascadeCustomHeaderWithRowSelectionActionEnabled: StoryObj<
+  { query: string } & Pick<ComponentProps<typeof DataCascade>, 'size'>
+> = {
+  render: function DataCascadeWrapper(args, context) {
+    const groupByFields = useMemo(
+      () => getESQLStatsQueryMeta(args.query).groupByFields.map(({ field }) => field),
+      [args.query]
+    );
+
+    const generateGroupFieldRecord = useCallback(
+      (nodePath?: string[], nodePathMap?: Record<string, string>) => {
+        return groupByFields.reduce<Record<string, string>>((acc, field) => {
+          return {
+            ...acc,
+            [field]:
+              nodePathMap && nodePath?.indexOf(field) !== -1
+                ? nodePathMap[field]
+                : /clientip/.test(field)
+                ? faker.internet.ipv4()
+                : /url\.keyword/.test(field)
+                ? faker.internet.url({ protocol: 'https' })
+                : faker.person.fullName(),
+          };
+        }, {} as Record<string, string>);
+      },
+      [groupByFields]
+    );
+
+    const initData: MockGroupData[] = useMemo(
+      () =>
+        new Array(100).fill(null).map(() => {
+          return {
+            id: faker.string.uuid(),
+            count: faker.number.int({ min: 1, max: 100 }),
+            customer_email: faker.internet.email(),
+            ...generateGroupFieldRecord(),
+          };
+        }),
+      [generateGroupFieldRecord]
+    );
+
+    const customTableHeader = useCallback<
+      NonNullable<ComponentProps<typeof DataCascade<MockGroupData>>['customTableHeader']>
+    >(({ currentSelectedColumns, availableColumns, onGroupSelection, selectedRows }) => {
+      return (
+        <EuiFlexGroup justifyContent="spaceBetween" alignItems="center">
+          <EuiFlexItem grow={false}>
+            <EuiFlexGroup alignItems="center" gutterSize="s">
+              <EuiFlexItem grow={false}>
+                <EuiIcon type="securitySignal" size="xl" aria-hidden={true} />
+              </EuiFlexItem>
+              <EuiFlexItem grow={true}>
+                <EuiText>
+                  <h2>Security alerts</h2>
+                </EuiText>
+              </EuiFlexItem>
+            </EuiFlexGroup>
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiFlexGroup justifyContent="center" alignItems="center" gutterSize="s">
+              {selectedRows.length > 0 && (
+                <EuiFlexItem alignItems="center" gutterSize="s">
+                  <EuiFilterGroup>
+                    <EuiFilterButton
+                      iconType="arrowDown"
+                      onClick={() => {
+                        // Handle filter button click
+                      }}
+                      badgeColor="success"
+                      hasActiveFilters={true}
+                      numFilters={selectedRows.length}
+                    >
+                      Selected
+                    </EuiFilterButton>
+                  </EuiFilterGroup>
+                </EuiFlexItem>
+              )}
+              <EuiFlexItem grow={false}>
+                <EuiText size="s" color="subdued">
+                  <strong>Group by:</strong>
+                </EuiText>
+              </EuiFlexItem>
+              <EuiFlexItem>
+                <EuiButtonGroup
+                  legend="select columns"
+                  idSelected={currentSelectedColumns[0]}
+                  options={availableColumns.map((col) => ({ id: col, label: col }))}
+                  onChange={(id) => {
+                    onGroupSelection([id]);
+                  }}
+                />
+              </EuiFlexItem>
+            </EuiFlexGroup>
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      );
+    }, []);
+
+    const onCascadeGroupingChange = useCallback<
+      NonNullable<ComponentProps<typeof DataCascade<MockGroupData>>['onCascadeGroupingChange']>
+    >((groupBy) => {
+      // eslint-disable-next-line no-console -- Handle group by change if needed
+      console.log('Group By Changed:', groupBy);
+    }, []);
+
+    const onCascadeGroupNodeExpanded = useCallback<
+      DataCascadeRowProps<MockGroupData, LeafNode>['onCascadeGroupNodeExpanded']
+    >(
+      async ({ row, nodePath, nodePathMap }) => {
+        // Simulate a data fetch on row expansion
+        return new Promise((resolve) => {
+          setTimeout(() => {
+            resolve(
+              new Array(row.count).fill(null).map(() => {
+                return {
+                  id: faker.string.uuid(),
+                  count: faker.number.int({ min: 1, max: row.count }),
+                  ...generateGroupFieldRecord(nodePath, nodePathMap),
+                };
+              })
+            );
+          }, 3000);
+        });
+      },
+      [generateGroupFieldRecord]
+    );
+
+    const rowHeaderTitleSlot = useCallback<
+      DataCascadeRowProps<MockGroupData, LeafNode>['rowHeaderTitleSlot']
+    >(({ rowData, nodePath }) => {
+      const rowGroup = nodePath[nodePath.length - 1];
+
+      return (
+        <EuiText>
+          <h2>{rowData[rowGroup]}</h2>
+        </EuiText>
+      );
+    }, []);
+
+    const rowHeaderMetaSlots = useCallback<
+      NonNullable<DataCascadeRowProps<MockGroupData, LeafNode>['rowHeaderMetaSlots']>
+    >(({ rowData }) => {
+      return [
+        <EuiStat
+          reverse
+          title={rowData.count}
+          textAlign="right"
+          description={
+            <FormattedMessage
+              id="sharedUXPackages.data_cascade.demo.row.count"
+              defaultMessage="<indicator>record count</indicator>"
+              values={{
+                indicator: (chunks) => <EuiHealth color="subdued">{chunks}</EuiHealth>,
+              }}
+            />
+          }
+        />,
+      ];
+    }, []);
+
+    const onCascadeLeafNodeExpanded = useCallback<
+      DataCascadeRowCellProps<MockGroupData, LeafNode>['onCascadeLeafNodeExpanded']
+    >(
+      async ({ row, nodePathMap, nodePath }) => {
+        // Simulate a data fetch for the expanded leaf,
+        // ideally we'd want to use nodePath information to fetch this data
+        return new Promise<LeafNode[]>((resolve) => {
+          setTimeout(() => {
+            resolve(
+              new Array(row.count).fill(null).map(() => {
+                return {
+                  id: faker.string.uuid(),
+                  ...generateGroupFieldRecord(nodePath, nodePathMap),
+                };
+              })
+            );
+          }, 3000);
+        });
+      },
+      [generateGroupFieldRecord]
+    );
+
+    const cascadeLeafCellRenderer = useCallback<
+      DataCascadeRowCellProps<MockGroupData, LeafNode>['children']
+    >(
+      ({ data }) => {
+        return (
+          <EuiBasicTable
+            tableCaption="custom header with row selection action enabled table"
+            columns={[
+              {
+                field: 'id',
+                name: 'ID',
+              },
+              ...groupByFields.map((field, index, groupArray) => ({
+                field,
+                name: field.replace(/_/g, ' '),
+                ...(index === groupArray.length - 1
+                  ? { align: 'right' as HorizontalAlignment }
+                  : {}),
+              })),
+            ]}
+            items={(data ?? []).map((datum) => ({
+              id: datum.id,
+              ...groupByFields.reduce(
+                (acc, field) => ({
+                  ...acc,
+                  [field]: datum[field],
+                }),
+                {} as Record<string, string>
+              ),
+            }))}
+          />
+        );
+      },
+      [groupByFields]
+    );
+
+    const rowHeaderActions = useCallback<
+      NonNullable<DataCascadeRowProps<MockGroupData, LeafNode>['rowHeaderActions']>
+    >(({ rowData, nodePath }) => {
+      const rowGroup = nodePath[nodePath.length - 1];
+
+      const groupValue = rowData[rowGroup];
+
+      return [
+        {
+          iconType: 'reporter',
+          'aria-label': `investigate ${groupValue}`,
+          onClick: () => {
+            /** Noop Click handler for starting an investigation */
+          },
+          label: (
+            <FormattedMessage
+              id="sharedUXPackages.data_cascade.demo.row.investigate"
+              defaultMessage="Investigate"
+            />
+          ),
+        },
+        {
+          iconType: 'flag',
+          'aria-label': `flag ${groupValue}`,
+          onClick: () => {
+            /** Noop Click handler for flagging */
+          },
+          label: (
+            <FormattedMessage
+              id="sharedUXPackages.data_cascade.demo.row.flag"
+              defaultMessage="Flag"
+            />
+          ),
+        },
+        {
+          iconType: 'bell',
+          'aria-label': `Create alert for ${groupValue}`,
+          onClick: () => {
+            /** Noop Click handler for create alerts */
+          },
+          label: (
+            <FormattedMessage
+              id="sharedUXPackages.data_cascade.demo.row.create_alerts"
+              defaultMessage="Create alerts"
+            />
+          ),
+        },
+        {
+          iconType: 'download',
+          'aria-label': `Download records of ${groupValue}`,
+          onClick: () => {
+            /** Noop Click handler for download */
+          },
+          label: (
+            <FormattedMessage
+              id="sharedUXPackages.data_cascade.demo.row.download"
+              defaultMessage="Download"
+            />
+          ),
+        },
+      ];
+    }, []);
+
+    return (
+      <EuiFlexGroup direction="column" css={{ height: 'calc(100svh - 2rem)' }}>
+        <EuiFlexItem grow={false}>
+          <EuiText>
+            <div>
+              <h3>{context.name}</h3>
+              <p>Data Cascade with row selection action enabled</p>
+            </div>
+          </EuiText>
+        </EuiFlexItem>
+        <EuiFlexItem>
+          <DataCascade
+            size={args.size}
+            data={initData}
+            cascadeGroups={groupByFields}
+            customTableHeader={customTableHeader}
+            onCascadeGroupingChange={onCascadeGroupingChange}
+            enableRowSelection
+          >
+            <DataCascadeRow<MockGroupData, LeafNode>
+              onCascadeGroupNodeExpanded={onCascadeGroupNodeExpanded}
+              rowHeaderTitleSlot={rowHeaderTitleSlot}
+              rowHeaderMetaSlots={rowHeaderMetaSlots}
+              rowHeaderActions={rowHeaderActions}
+            >
+              <DataCascadeRowCell onCascadeLeafNodeExpanded={onCascadeLeafNodeExpanded}>
+                {cascadeLeafCellRenderer}
+              </DataCascadeRowCell>
+            </DataCascadeRow>
+          </DataCascade>
+        </EuiFlexItem>
+      </EuiFlexGroup>
+    );
+  },
+  argTypes: {
+    query: {
+      type: 'string' as const,
+      description: 'Simulation of The ES|QL query that the user provided into the esql editor,',
+    },
+    size: {
+      control: 'radio',
+      options: ['s', 'm', 'l'],
+      description: 'Size of the cascade rows',
+    },
+  },
+  args: {
+    query: 'FROM kibana_sample_data_logs | STATS count = COUNT() BY clientip, url.keyword',
+    size: 'm',
+  },
+};
+
+CascadeCustomHeaderWithRowSelectionActionEnabled.parameters = { docs: { disable: true } };
